# api-docs.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "Katron AI Gift Card API Documentation",
                 "version":  "1.0"
             },
    "servers":  [
                    {
                        "url":  "https://api.ktngiftcard.katronai.com/katron-gift-card",
                        "description":  "Generated server url"
                    }
                ],
    "security":  [
                     {
                         "BearerAuth":  [

                                        ]
                     }
                 ],
    "paths":  {
...
                  "/api/giftCards/purchaseGiftCardByMerchant":  {
                                                                    "post":  {
                                                                                 "tags":  [
                                                                                              "gift-card-controller"
                                                                                          ],
                                                                                 "operationId":  "purchaseGiftCardByMerchant",
                                                                                 "requestBody":  {
                                                                                                     "content":  {
                                                                                                                     "application/json":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/GiftcardOrderRequestMerchant"
                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                 },
                                                                                                     "required":  true
                                                                                                 },
                                                                                 "responses":  {
                                                                                                   "200":  {
                                                                                                               "description":  "OK",
                                                                                                               "content":  {
                                                                                                                               "*/*":  {
                                                                                                                                           "schema":  {
                                                                                                                                                          "$ref":  "#/components/schemas/Response"
                                                                                                                                                      }
                                                                                                                                       }
                                                                                                                           }
                                                                                                           }
                                                                                               }
                                                                             }
                                                                },
...
                  "/api/giftCards/checkForPaymentAndGiftCardStatus":  {
                                                                          "post":  {
                                                                                       "tags":  [
                                                                                                    "gift-card-controller"
                                                                                                ],
                                                                                       "operationId":  "checkForPaymentAndGiftCardStatus",
                                                                                       "parameters":  [
                                                                                                          {
                                                                                                              "name":  "giftCardOrderId",
                                                                                                              "in":  "query",
                                                                                                              "required":  true,
                                                                                                              "schema":  {
                                                                                                                             "type":  "integer",
                                                                                                                             "format":  "int64"
                                                                                                                         }
                                                                                                          }
                                                                                                      ],
                                                                                       "responses":  {
                                                                                                         "200":  {
                                                                                                                     "description":  "OK",
                                                                                                                     "content":  {
                                                                                                                                     "*/*":  {
                                                                                                                                                 "schema":  {
                                                                                                                                                                "$ref":  "#/components/schemas/Response"
                                                                                                                                                            }
                                                                                                                                             }
                                                                                                                                 }
                                                                                                                 }
                                                                                                     }
                                                                                   }
                                                                      },
...
                  "/api/external/giftCard/resendGiftCardCredentials":  {
                                                                           "post":  {
                                                                                        "tags":  [
                                                                                                     "external-customer-controller"
                                                                                                 ],
                                                                                        "operationId":  "resendGiftCardCredentials",
                                                                                        "parameters":  [
                                                                                                           {
                                                                                                               "name":  "giftCardOrderId",
                                                                                                               "in":  "query",
                                                                                                               "required":  true,
                                                                                                               "schema":  {
                                                                                                                              "type":  "integer",
                                                                                                                              "format":  "int64"
                                                                                                                          }
                                                                                                           }
                                                                                                       ],
                                                                                        "responses":  {
                                                                                                          "200":  {
                                                                                                                      "description":  "OK",
                                                                                                                      "content":  {
                                                                                                                                      "*/*":  {
                                                                                                                                                  "schema":  {
                                                                                                                                                                 "$ref":  "#/components/schemas/Response"
                                                                                                                                                             }
                                                                                                                                              }
                                                                                                                                  }
                                                                                                                  }
                                                                                                      }
                                                                                    }
                                                                       },
...
                  "/api/giftCards/listAllGiftCards":  {
                                                          "get":  {
                                                                      "tags":  [
                                                                                   "gift-card-controller"
                                                                               ],
                                                                      "operationId":  "listAllGiftCards",
                                                                      "parameters":  [
                                                                                         {
                                                                                             "name":  "type",
                                                                                             "in":  "query",
                                                                                             "required":  true,
                                                                                             "schema":  {
                                                                                                            "type":  "string"
                                                                                                        }
                                                                                         }
                                                                                     ],
                                                                      "responses":  {
                                                                                        "200":  {
                                                                                                    "description":  "OK",
                                                                                                    "content":  {
                                                                                                                    "*/*":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/Response"
                                                                                                                                           }
                                                                                                                            }
                                                                                                                }
                                                                                                }
                                                                                    }
                                                                  }
                                                      },
...
              },
    "components":  {
                       "schemas":  {
                                       "Response":  {
                                                        "type":  "object",
                                                        "properties":  {
                                                                           "status":  {
                                                                                          "type":  "integer",
                                                                                          "format":  "int32"
                                                                                      },
                                                                           "message":  {
                                                                                           "type":  "string"
                                                                                       },
                                                                           "data":  {
                                                                                        "type":  "object"
                                                                                    }
                                                                       }
                                                    },
...
                                       "GiftcardOrderRequestMerchant":  {
                                                                            "required":  [
                                                                                             "email",
                                                                                             "name"
                                                                                         ],
                                                                            "type":  "object",
                                                                            "properties":  {
                                                                                               "giftCardId":  {
                                                                                                                  "type":  "integer",
                                                                                                                  "format":  "int64"
                                                                                                              },
                                                                                               "quantity":  {
                                                                                                                "type":  "integer",
                                                                                                                "format":  "int32"
                                                                                                            },
                                                                                               "unitPrice":  {
                                                                                                                 "type":  "number",
                                                                                                                 "format":  "float"
                                                                                                             },
                                                                                               "email":  {
                                                                                                             "type":  "string"
                                                                                                         },
                                                                                               "name":  {
                                                                                                            "type":  "string"
                                                                                                        }
                                                                                           }
                                                                        },
...
                                   },
                       "securitySchemes":  {
                                               "BearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}